+++ tests/test_file.py
@@ -19,35 +19,43 @@ class FileTest(unittest.TestCase):
         os.environ["ISITWEEKENDYET"] = "0"
         self.assertFalse(psf.env_get("ISITWEEKENDYET", boolean=True))
 
-    def test_get_sentinel_scene_from_dir(self):
+    def test_get_sentinel_scene_from_path_testfiles(self):
         with psf.get_sentinel_scene_from_dir(path_testfiles) as (full_path, ident):
             self.assertEqual("S1M_hello_from_inside", ident)
 
+    def test_get_sentinel_scene_from_str_path(self):
         with psf.get_sentinel_scene_from_dir(str_path) as (full_path, ident):
             self.assertEqual("S1M_hello_from_inside", ident)
 
+    def test_get_sentinel_scene_from_dir_joinpath(self):
         with psf.get_sentinel_scene_from_dir(path_testfiles.joinpath("another_scene")) as (full_path, ident):
             self.assertEqual("S2__IN_FOLDER", ident)
 
-    def test_get_polarization_from_s1_filename(self):
+    def test_get_polarization_from_s1_filename_SDH(self):
         self.assertEqual(
             psf.get_polarization_from_s1_filename(
                 "MMM_BB_TTTR_1SDH_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_OOOOOO_DDDDDD_CCCC.SAFE.zip"
             ),
             "HH",
         )
+
+    def test_get_polarization_from_s1_filename_SSH(self):
         self.assertEqual(
             psf.get_polarization_from_s1_filename(
                 "MMM_BB_TTTR_1SSH_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_OOOOOO_DDDDDD_CCCC.SAFE.zip"
             ),
             "HH",
         )
+
+    def test_get_polarization_from_s1_filename_SSV(self):
         self.assertEqual(
             psf.get_polarization_from_s1_filename(
                 "MMM_BB_TTTR_2SSV_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_OOOOOO_DDDDDD_CCCC.SAFE.zip"
             ),
             "VV",
         )
+
+    def test_get_polarization_from_s1_filename_SDV(self):
         self.assertEqual(
             psf.get_polarization_from_s1_filename(
                 "MMM_BB_TTTR_1SDV_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_OOOOOO_DDDDDD_CCCC.SAFE.zip",
@@ -56,13 +64,7 @@ class FileTest(unittest.TestCase):
             "VV,VH",
         )
 
-    def test_get_ts_from_sentinel_filename(self):
-        self.assertEqual(
-            psf.get_ts_from_sentinel_filename(
-                "S3M_OL_L_TTT____20200113T002219_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_i_GGG_c.SEN3"
-            ),
-            datetime(2020, 1, 13, 0, 22, 19, tzinfo=timezone.utc),
-        )
+    def test_get_ts_from_sentinel_filename_S1(self):
         self.assertEqual(
             psf.get_ts_from_sentinel_filename(
                 "S1M_BB_TTTR_LFPP_YYYYMMDDTHHMMSS_20200113T002219_OOOOOO_DDDDDD_CCCC.SAFE.zip",
@@ -76,9 +78,19 @@ class FileTest(unittest.TestCase):
             ),
             datetime(2020, 1, 13, 7, 46, 19, tzinfo=timezone.utc),
         )
+
+    def test_get_ts_from_sentinel_filename_S2(self):
         self.assertEqual(
             psf.get_ts_from_sentinel_filename(
                 "S2AM_MSIXXX_20200113T002219_Nxxyy_ROOO_Txxxxx_<Product Discriminator>.SAFE"
+            ),
+            datetime(2020, 1, 13, 0, 22, 19, tzinfo=timezone.utc),
+        )
+
+    def test_get_ts_from_sentinel_filename_S3(self):
+        self.assertEqual(
+            psf.get_ts_from_sentinel_filename(
+                "S3M_OL_L_TTT____20200113T002219_YYYYMMDDTHHMMSS_YYYYMMDDTHHMMSS_i_GGG_c.SEN3"
             ),
             datetime(2020, 1, 13, 0, 22, 19, tzinfo=timezone.utc),
         )