+++ tests/test_raster.py
@@ -49,9 +49,8 @@ class RasterTest(unittest.TestCase):
             Image(data=self.img.arr, crs=self.img.dataset.crs)
 
     def test_init_with_array(self):
-        img = Image(self.img.arr, crs=self.img.dataset.crs, transform=self.img.dataset.transform)
-        self.assertTrue(np.array_equal(self.img.arr, img.arr))
-        img.close()
+        with Image(self.img.arr, crs=self.img.dataset.crs, transform=self.img.dataset.transform) as img:
+            self.assertTrue(np.array_equal(self.img.arr, img.arr))
 
     def test_init_2dim(self):
         with Image(np.ones(shape=(385, 502)), crs=self.img.dataset.crs, transform=self.img.dataset.transform) as img:
@@ -67,68 +66,67 @@ class RasterTest(unittest.TestCase):
         with self.assertRaises(TypeError):
             Image(TEST_FILE, dimorder="middle")
 
-    def test_arr(self):
-        img_first = Image(TEST_FILE, dimorder="first")
-        img_first.mask(
-            box(
-                11.9027457562112939,
-                51.4664152338322580,
-                11.9477435281016131,
-                51.5009522690838750,
-            )
-        )
-        self.assertEqual(img_first.arr.shape, (1, 385, 502))
-        self.assertEqual(
-            str(img_first.dataset.transform),
-            str(
-                from_bounds(
+    def test_file_dimorder_first(self):
+        with Image(TEST_FILE, dimorder="first") as img_first:
+            img_first.mask(
+                box(
                     11.9027457562112939,
                     51.4664152338322580,
                     11.9477435281016131,
                     51.5009522690838750,
-                    img_first.arr.shape[2],
-                    img_first.arr.shape[1],
                 )
-            ),
-        )
-        img_first.close()
-
-        img_last = Image(TEST_FILE, dimorder="last")
-        img_last.mask(
-            box(
-                11.9027457562112939,
-                51.4664152338322580,
-                11.9477435281016131,
-                51.5009522690838750,
             )
-        )
-        self.assertEqual(img_last.arr.shape, (385, 502, 1))
-        self.assertEqual(
-            str(img_last.dataset.transform),
-            str(
-                from_bounds(
+            self.assertEqual(img_first.arr.shape, (1, 385, 502))
+            self.assertEqual(
+                str(img_first.dataset.transform),
+                str(
+                    from_bounds(
+                        11.9027457562112939,
+                        51.4664152338322580,
+                        11.9477435281016131,
+                        51.5009522690838750,
+                        img_first.arr.shape[2],
+                        img_first.arr.shape[1],
+                    )
+                ),
+            )
+
+    def test_file_dimorder_last(self):
+        with Image(TEST_FILE, dimorder="last") as img_last:
+            img_last.mask(
+                box(
                     11.9027457562112939,
                     51.4664152338322580,
                     11.9477435281016131,
                     51.5009522690838750,
-                    img_last.arr.shape[1],
-                    img_last.arr.shape[0],
                 )
-            ),
-        )
-        img_last.close()
+            )
+            self.assertEqual(img_last.arr.shape, (385, 502, 1))
+            self.assertEqual(
+                str(img_last.dataset.transform),
+                str(
+                    from_bounds(
+                        11.9027457562112939,
+                        51.4664152338322580,
+                        11.9477435281016131,
+                        51.5009522690838750,
+                        img_last.arr.shape[1],
+                        img_last.arr.shape[0],
+                    )
+                ),
+            )
 
-        img_first = Image(
+    def test_arr_dimorder_first(self):
+        with Image(
             np.ones((1, 385, 502)), dimorder="first", crs=self.img.dataset.crs, transform=self.img.dataset.transform
-        )
-        self.assertEqual(img_first.arr.shape, (1, 385, 502))
-        img_first.close()
+        ) as img_first:
+            self.assertEqual(img_first.arr.shape, (1, 385, 502))
 
-        img_last = Image(
+    def test_arr_dimorder_last(self):
+        with Image(
             np.ones((385, 502, 1)), dimorder="last", crs=self.img.dataset.crs, transform=self.img.dataset.transform
-        )
-        self.assertEqual(img_last.arr.shape, (385, 502, 1))
-        img_last.close()
+        ) as img_last:
+            self.assertEqual(img_last.arr.shape, (385, 502, 1))
 
     def test_arr_nodata(self):
         array = np.ones((3, 385, 502))
@@ -144,7 +142,7 @@ class RasterTest(unittest.TestCase):
 
             self.assertTrue(np.array_equal(im.arr, np.ones(shape=im.arr.shape)))
 
-    def test_set_array_error(self):
+    def test_set_array_error_dimorder_first(self):
         img_first = Image(TEST_FILE, dimorder="first")
 
         with self.assertRaises(TypeError):
@@ -153,6 +151,7 @@ class RasterTest(unittest.TestCase):
         with self.assertRaises(ValueError):
             img_first.arr = np.ones(shape=(764, 679, 1))
 
+    def test_set_array_error_dimorder_last(self):
         with self.assertRaises(ValueError):
             img_last = Image(TEST_FILE, dimorder="last")
             img_last.arr = np.ones(shape=(1, 764, 679))