+++ CHANGELOG.rst
@@ -16,12 +16,15 @@ something was changed
 [master] (2021-**-**)
 ---------------------
 
+Changed
+^^^^^^^
+- ``data``: platform removed from download_image and download_quicklook #130
+
 [1.1.0] (2021-03-29)
 ---------------------
 
 Changed
 ^^^^^^^
-- ``data``: platform removed from download_image and download_quicklook #130
 - ``data``: prep_aoi() is not private and supports __geo_interface__ #113
 - ``data``: change default scihup endpoint `/dhus` to `/apihub` #110
 - ``data``: mocked all interaction with Hubs for unit testing #116