+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Changed
 
 Changed
 ^^^^^^^
+- ``data``: generator added to query_metadata and query_metadata_srcid #127
 - ``data``: prep_aoi() is not private and supports __geo_interface__ #113
 - ``data``: change default scihup endpoint `/dhus` to `/apihub` #110
 - ``data``: mocked all interaction with Hubs for unit testing #116